# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Sober app",
        "title": "Sober API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/drink-logs": {
            "get": {
                "description": "Retrieve all drink logs for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drinks"
                ],
                "summary": "Get drink logs for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.DrinkLogsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new drink log for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drinks"
                ],
                "summary": "Create a drink log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create drink log request",
                        "name": "drinkLog",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateDrinkLogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateDrinkLogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/drink-logs/parse": {
            "post": {
                "description": "Parse a drink log and return the drink option and confidence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drinks"
                ],
                "summary": "Parse a drink log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Parse drink log request",
                        "name": "drinkLog",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ParseDrinkLogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ParseDrinkLogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/drink-options": {
            "get": {
                "description": "Retrieve all drink options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drinks"
                ],
                "summary": "Get all drink options",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.DrinkOptionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/drink-options/{id}": {
            "get": {
                "description": "Retrieve a specific drink option by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drinks"
                ],
                "summary": "Get a specific drink option",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drink option ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.DrinkOptionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a specific drink option by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drinks"
                ],
                "summary": "Update a drink option",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drink option ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated drink option",
                        "name": "drinkOption",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateDrinkOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific drink option by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drinks"
                ],
                "summary": "Delete a drink option",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Drink option ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Similar to the http status code",
                    "type": "integer",
                    "x-order": "1",
                    "example": 400
                },
                "type": {
                    "description": "The type of error",
                    "type": "string",
                    "enum": [
                        "validation",
                        "database",
                        "entity"
                    ],
                    "x-order": "2",
                    "example": "validation"
                },
                "message": {
                    "description": "A human-readable error message",
                    "type": "string",
                    "x-order": "3",
                    "example": "Invalid request Body"
                },
                "details": {
                    "description": "Additional details about the error, omitted if empty",
                    "type": "array",
                    "items": {
                        "type": "object"
                    },
                    "x-order": "4"
                },
                "correlation_id": {
                    "description": "The error id from context traceId",
                    "type": "string",
                    "x-order": "5",
                    "example": "01234567890123456789012345678900"
                }
            }
        },
...
        "dtos.CreateDrinkLogRequest": {
            "type": "object",
            "properties": {
                "drink_option_id": {
                    "type": "integer"
                },
                "logged_at": {
                    "type": "string"
                }
            }
        },
        "dtos.CreateDrinkLogResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "dtos.DrinkLogsResponse": {
            "type": "object",
            "properties": {
                "drink_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DrinkLog"
                    }
                }
            }
        },
        "dtos.DrinkOptionResponse": {
            "type": "object",
            "properties": {
                "drink_option": {
                    "$ref": "#/definitions/models.DrinkOption"
                }
            }
        },
        "dtos.DrinkOptionsResponse": {
            "type": "object",
            "properties": {
                "drink_options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DrinkOption"
                    }
                }
            }
        },
        "dtos.ParseDrinkLogRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "dtos.ParseDrinkLogResponse": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "drink_option": {
                    "$ref": "#/definitions/models.DrinkOption"
                }
            }
        },
        "dtos.UpdateDrinkOptionRequest": {
            "type": "object",
            "required": [
                "abv",
                "name",
                "size_unit",
                "size_value",
                "type"
            ],
            "properties": {
                "abv": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "size_unit": {
                    "type": "string"
                },
                "size_value": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.DrinkLog": {
            "type": "object",
            "properties": {
                "abv": {
                    "type": "number"
                },
                "drink_name": {
                    "type": "string"
                },
                "drink_option_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "logged_at": {
                    "type": "string"
                },
                "size_unit": {
                    "type": "string"
                },
                "size_value": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.DrinkOption": {
            "type": "object",
            "properties": {
                "abv": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "size_unit": {
                    "type": "string"
                },
                "size_value": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}